nothing to see here

+++ core/src/main/java/com/krolewskie_potyczki/model/config/EnemyConstantsConfig.java
@@ -1,7 +1,7 @@
 package com.krolewskie_potyczki.model.config;
 
 public class EnemyConstantsConfig {
-    public int basicRandomEnemySpawn;
+    public float basicRandomEnemySpawn;
 
     public EnemyConstantsConfig() {}
 }
